refactor: replaced interface with domain entity
